requirement.txt
opencv-python === 4.8.1.78
selenium === 4.14.0
openpyxl === 3.0.4
lxml === 4.9.3
pillow === 9.5.0